% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnds_across_species_plot.R
\name{dnds_across_multiple_species_plot}
\alias{dnds_across_multiple_species_plot}
\title{Plot pairwise dN, dS, or dN/dS distributions of orthologous genes}
\usage{
dnds_across_multiple_species_plot(
  dnds_species_tbl,
  type,
  order,
  xlab = "",
  ylab = "",
  title = ""
)
}
\arguments{
\item{dnds_species_tbl}{a dnds table created with \code{\link{dnds_across_multiple_species}}.}

\item{type}{type of substitution rate quantification that shall be visualized on the y-axis. Options are:
\itemize{
\item \code{type = "dn"}
\item \code{type = "ds"}
\item \code{type = "dn/ds"}
\item \code{type = "dn+ds"}
}}

\item{order}{a character vector specifying the order of the species using their scientific names (which have to match the species names stored in the \code{subject_species}
column of the input \code{dnds_species_tbl}).}

\item{xlab}{label of x-axis.}

\item{ylab}{label of y-axis.}

\item{title}{title label.}
}
\description{
This function enables users to plot pairwise dN, dS, or dN/dS distributions of orthologous genes.
}
\author{
Hajk-Georg Drost
}
