% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cds.R
\name{import_cds}
\alias{import_cds}
\title{Import the coding sequences from a fasta file}
\usage{
import_cds(file, format, delete_corrupt_cds = TRUE, ...)
}
\arguments{
\item{file}{a character string specifying the path to the file storing the CDS.}

\item{format}{a character string specifying the file format used to store the CDS, e.g. "fasta", "fatsq".}

\item{delete_corrupt_cds}{a logical value indicating whether sequences with corrupt base triplets should be removed from the input \code{file}. This is the case when the length of coding sequences cannot be divided by 3 and thus the coding sequence contains at least one corrupt base triplet.}

\item{...}{additional arguments that are used by the \code{\link[Biostrings]{readDNAStringSet}} function.}
}
\value{
A data.table storing the gene id in the first column and the corresponding
sequence as string in the second column.
}
\description{
This function reads an organism specific CDS stored in a defined file format.
}
\details{
The \code{import_cds} function takes a string specifying the path to the cds file
of interest as first argument.

It is possible to read in different proteome file standards such as \emph{fasta} or \emph{fastq}.

CDS stored in fasta files can be downloaded from http://www.ensembl.org/info/data/ftp/index.html.
}
\examples{
\dontrun{
# reading a cds file stored in fasta format
Ath_cds <- import_cds(system.file('seqs/ortho_thal_cds.fasta', package = 'homologr'),
                    format = "fasta")
# look at results
Ath_cds
}

}
\author{
Hajk-Georg Drost
}
