% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dnds_tbl.R
\name{import_dnds_tbl}
\alias{import_dnds_tbl}
\title{Import a dnds table generated with \code{\link{dnds}}}
\usage{
import_dnds_tbl(file, sep = ";")
}
\arguments{
\item{file}{file path to dnds table with \code{sep} separated columns.}

\item{sep}{column separator used in the input \code{\link{dnds}} table.}
}
\description{
This function reads a file that stores a dnds table
generated with \code{\link{dnds}}.
}
\examples{
\dontrun{
# generate dnds table
dNdS_tbl <- dnds(query_file      = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
subject_file    = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
ortho_detection = "DIAMOND-RBH",
aa_aln_type     = "pairwise",
aa_aln_tool     = "NW",
codon_aln_tool  = "pal2nal",
dnds_estimation = "Li",
cores           = 1 )

# save dnds table as ';' column separated file
utils::write.table(
dNdS_tbl,
file.path(tempdir(), "dNdS_tbl.csv"), sep = ";",
col.names = TRUE,
row.names = FALSE,
quote     = FALSE )

# import dNdS table into R session
dNdS_tbl_import <- import_dnds_tbl(file.path(tempdir(), "dNdS_tbl.csv"))
}
}
\author{
Hajk-Georg Drost
}
