% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_multispecies_homology_between_genes.R
\name{infer_multispecies_homology_between_genes}
\alias{infer_multispecies_homology_between_genes}
\title{Multispecies Homology Inference of Genes}
\usage{
infer_multispecies_homology_between_genes()
}
\description{
Multispecies Homology Inference of Genes
}
